+++ nf_core/modules/__init__.py
@@ -1,0 +1,3 @@
+from .pipeline_modules import ModulesRepo, PipelineModules
+from .create import ModuleCreate
+from .test_yml_builder import ModulesTestYmlBuilder
